backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
openai-whisper
torch>=2.0.0
numpy
pydantic==2.6.1
python-dotenv==1.0.1

